feat: 최근 리포트 트렌드 API 추가 및 스토리 재-import 버그 수정
- GET /v1/story-completions/recent 추가 - 최근 N회(최대 20) 리포트를 outcomes와
  함께 반환해 프론트의 누적 접근 방식 트렌드 카드가 여러 회차를 가로질러 계산할 수 있게 함
- StoryActionFamilyRepository.deleteByAnchor_IdInAndOrigin의 반환 타입을 long에서
  int로 수정 - Spring Data JPA의 @Modifying 쿼리는 void/int/Integer만 허용하는데
  long을 써서 호출될 때마다(=스토리 재-import 시마다) IllegalArgumentException이
  나던 버그

+++ src/main/java/com/qstory/backend/story/repository/StoryActionFamilyRepository.java
@@ -23,7 +23,7 @@ public interface StoryActionFamilyRepository extends JpaRepository<StoryActionFa
      */
     @Modifying
     @Query("delete from StoryActionFamily f where f.anchor.id in :anchorIds and f.origin = :origin")
-    long deleteByAnchor_IdInAndOrigin(@Param("anchorIds") List<String> anchorIds, @Param("origin") FamilyOrigin origin);
+    int deleteByAnchor_IdInAndOrigin(@Param("anchorIds") List<String> anchorIds, @Param("origin") FamilyOrigin origin);
 
     /** LiveBranchGenerationService.enqueue()의 앵커당 상한 체크. */
     long countByAnchor_IdAndOrigin(String anchorId, FamilyOrigin origin);
